trip_planner_www/requirements.txt
Django==1.6.1
jsonfield
django_admin_bootstrapped
django-geoip
django-tastypie
psycopg2==2.6.1